Odfeature/mic check

+++ app/channels/appearances_channel.rb
@@ -4,6 +4,17 @@ class AppearancesChannel < ApplicationCable::Channel
   end
 
   def unsubscribed
-    guest.destroy
+    guest.destroy if guest
+  end
+
+  def update(data)
+    puts "Updating guest status..."
+    return unless guest
+    if guest.update_attributes(status: data['status'])
+      ActionCable.server.broadcast("appearances_#{params[:lobby]}",
+                                   guest: guest.name,
+                                   guest_id: guest.id,
+                                   status: guest.status)
+    end
   end
 end
